fix wrangler
