+++ src/amr/messengers/messenger_factory.h
@@ -27,20 +27,9 @@ namespace amr
 
 
 
-    std::vector<MessengerDescriptor> makeDescriptors(std::vector<std::string> modelNames)
-    {
-        if (modelNames.size() == 1)
-            return {{modelNames[0], modelNames[0]}};
+    std::vector<MessengerDescriptor> makeDescriptors(std::vector<std::string> modelNames);
 
-        else if (modelNames.size() == 2)
-        {
-            return {{modelNames[0], modelNames[0]},
-                    {modelNames[0], modelNames[1]},
-                    {modelNames[1], modelNames[1]}};
-        }
-        else
-            throw std::runtime_error("Error max number of models is 2");
-    }
+
 
 
     template<typename MHDModel, typename HybridModel, typename IPhysicalModel,
