+++ src/core/CMakeLists.txt
@@ -53,13 +53,16 @@ set( SOURCES_CPP
      data/ndarray/ndarray_vector.cpp
      data/ions/particle_initializers/maxwellian_particle_initializer.cpp
      utilities/index/index.cpp
+     utilities/mpi_utils.cpp
     )
+
+find_package(MPI)
 
 add_library(${PROJECT_NAME}  ${SOURCES_INC} ${SOURCES_CPP})
 target_compile_options(${PROJECT_NAME}  PRIVATE ${PHARE_WERROR_FLAGS})
-target_link_libraries(${PROJECT_NAME}  PRIVATE phare_initializer)
+target_link_libraries(${PROJECT_NAME}  PRIVATE phare_initializer ${MPI_C_LIBRARIES})
 
-target_include_directories(${PROJECT_NAME}  PUBLIC
+target_include_directories(${PROJECT_NAME}  PUBLIC ${MPI_C_INCLUDE_DIRS}
   $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/../../subprojects>)
